chore: refactor import statements for archive view widgets
