+++ .github/workflows/ci.yml
@@ -10,7 +10,7 @@ on:
 
 jobs:
   flutter_test:
-    name: Run flutter test and analyze
+    name: Run flutter test
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
