+++ lib/ui/archive/archive_view/widgets/build_idea_detail_page.dart
