+++ lib/ui/archive/archive_view/widgets/build_idea_tab_view.dart
@@ -1,7 +1,7 @@
 import 'package:classify/domain/models/memo/memo_model.dart';
 import 'package:flutter/material.dart';
 import 'package:classify/ui/archive/archive_view/view_models/archive_view_model.dart';
-import 'package:classify/ui/archive/archive_view/widgets/buildIdeaDetailPage.dart';
+import 'package:classify/ui/archive/archive_view/widgets/build_idea_detail_page.dart';
 import 'dart:math';
 import 'package:classify/utils/top_level_setting.dart';
