+++ lib/ui/archive/archive_view/widgets/build_study_detail_page.dart
